requirements.txt
annoy==1.17.3
datasets 
huggingface-hub==0.22.2
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.36
langchain-experimental==0.0.55
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.38
openai==1.12.0
python-dotenv 
streamlit==1.31.1
tiktoken==0.6.0
tokenizers==0.15.2
transformers==4.39.2
sentence-transformers

